naming changes, remove unused codes
