--- Tests/FeatherDatabaseTests/Mocks/MockDatabaseQuery.swift
@@ -1,13 +1,0 @@
-//
-//  MockDatabaseQuery.swift
-//  feather-database
-//
-//  Created by Tibor Bodecs on 2026. 01. 10..
-//
-
-import FeatherDatabase
-
-struct MockDatabaseQuery: DatabaseQuery {
-    let sql: String
-    let bindings: [String]
-}
